+++ public/sw.js
@@ -3,20 +3,15 @@
 const CACHE_NAME = 'finals-calculator-v8'
 
 const FILES_TO_CACHE = [
-  '/images/24x24.png',
-  '/images/32x32.png',
   '/images/48x48.png',
-  '/images/64x64.png',
+  '/images/72x72.png',
   '/images/96x96.png',
-  '/images/128x128.png',
-  '/images/256x256.png',
+  '/images/144x144.png',
   '/images/512x512.png',
-  '/images/1024x1024.png',
   '/images/bordered-512x512.png',
   '/images/maskable_icon.png',
   '/icon.png',
-  '/manifest/manifest-ko-KR.json',
-  '/manifest/manifest-en-US.json',
+  '/manifest.json',
   'https://1.www.s81c.com/common/carbon/plex/fonts/IBM-Plex-Sans/fonts/split/woff2/IBMPlexSans-Regular-Latin1.woff2',
   'https://1.www.s81c.com/common/carbon/plex/fonts/IBM-Plex-Sans/fonts/split/woff2/IBMPlexSans-Light-Latin1.woff2',
   'https://1.www.s81c.com/common/carbon/plex/fonts/IBM-Plex-Mono/fonts/split/woff2/IBMPlexMono-Regular-Latin1.woff2',