+++ CHANGELOG.md
@@ -6,3 +6,10 @@ Create basic calculator
 
 Make Rubiks Cube Modal
 Make Modal speed controllable
+
+# v0.0.3
+
+Make Rubiks Modal with dark outline
+Zoom in/out feature
+Add Total Move Count & Expected Time
+Fix jquery.cube.threejs.min.js bugs